REFACTOR: Move logic to only one class
